refactor: extension 위치 수정
